cast timeval.tv_usec to long
Problem: macos fails to compile some printfs of timeval.tv_usec
because it's an int on macos.

Cast where appropriate.

+++ src/cmd/builtin/dmesg.c
@@ -134,7 +134,7 @@ void print_iso_timestamp (struct dmesg_ctx *ctx, struct stdlog_header *hdr)
     printf ("%s%s.%.6lu%s%s ",
             dmesg_color (ctx, DMESG_COLOR_TIME),
             buf,
-            tv.tv_usec,
+            (unsigned long)tv.tv_usec,
             tz,
             dmesg_color_reset (ctx));
 }
